Tile ID & render changes
Tiles now change appearance when setID is called. Instead of passing Transform to a draw call you now have to pass a superclass of Transform - RenderStates, which gives the additional option to override the texture. Sprite::draw() was given a little optimization.

+++ src/transform.hpp
@@ -11,8 +11,5 @@ struct Transform {
 		return result;
 	}
 
-	glm::ivec2 translation;
-	// experimental
-	//double angle;
-
+	glm::vec2 translation;
 };
